+++ migrations/4_NonBurnablePausableInventory.js
@@ -1,5 +1,0 @@
-const NonBurnablePausableInventoryMock = artifacts.require('NonBurnablePausableInventoryMock');
-
-module.exports = async (deployer, network, accounts) => { 
-    await deployer.deploy(NonBurnablePausableInventoryMock, 2);
-}